New Features
Added USBhost input to the chaos() function. Made chaos() more chaotic. Added a 2 second timer on the enable pins so you can safely hotplug (but not hot-unplug foot controllers). Now, in edit mode, any number of things can be edited simultaniously. Fixed a bug that only allowed readAnalogMUX to toggle once. Got bulk send working. Disabled the pot on the side because it interferes with the expression pedal (sigh).